==> Course_Materials_Part2/Video_Lecture_NBs/NB_13_Groupby.ipynb <==
# Part 2: Full Data Workflow A-Z

## Pandas GroupBy Operations

### Understanding GroupBy objects

import pandas as pd

titanic = pd.read_csv("titanic.csv")

titanic.head()

titanic.tail()

titanic.info()

titanic_slice = titanic.iloc[:10, [2,3]]

titanic_slice

titanic_slice.groupby("sex")

gbo = titanic_slice.groupby("sex")

type(gbo)

gbo.groups

l = list(gbo)

l

len(l)

l[0]

type(l[0])

l[0][0]

l[0][1]

type(l[0][1])

l[1]

titanic_slice.loc[titanic_slice.sex == "female"]

titanic_slice_f = titanic_slice.loc[titanic_slice.sex == "female"]
titanic_slice_f

titanic_slice_m = titanic_slice.loc[titanic_slice.sex == "male"]
titanic_slice_m

titanic_slice_f.equals(l[0][1])

for element in gbo:
    print(element[1])

### Splitting with many Keys

import pandas as pd

summer = pd.read_csv("summer.csv")

summer.head()

summer.info()

summer.Country.nunique()

split1 = summer.groupby("Country")

l = list(split1)
l

len(l)

l[100][1]

split2 = summer.groupby(by = ["Country", "Gender"])

l2 = list(split2)
l2

len(l2)

l2[104]

l2[104][0]

l2[104][1]

### split-apply-combine explained

import pandas as pd

titanic = pd.read_csv("titanic.csv")

titanic_slice = titanic.iloc[:10, [2,3]]

titanic_slice

list(titanic_slice.groupby("sex"))[0][1]

list(titanic_slice.groupby("sex"))[1][1]

titanic_slice.groupby("sex").mean()

titanic.groupby("sex").survived.sum()

titanic.groupby("sex")[["fare", "age"]].max()

new_df = titanic.groupby("sex").mean(numeric_only = True)

new_df

import matplotlib.pyplot as plt
plt.style.use("seaborn-v0_8")

new_df.plot(kind = "bar", subplots = True, figsize = (8,15), fontsize = 13)
plt.show()

### split-apply-combine applied

import pandas as pd

summer = pd.read_csv("summer.csv")

summer.head()

summer.info()

medals_per_country = summer.groupby("Country").Medal.count().nlargest(n = 20)
medals_per_country

import matplotlib.pyplot as plt
plt.style.use("seaborn-v0_8")

medals_per_country.plot(kind = "bar", figsize = (14, 8), fontsize = 14)
plt.xlabel("Country", fontsize = 13)
plt.ylabel("No. of Medals", fontsize = 13)
plt.title("Summer Olympic Games (Total Medals per Country)", fontsize = 16)
plt.show()

titanic = pd.read_csv("titanic.csv")

titanic.head()

titanic.info()

titanic.describe()

titanic.fare.mean()

titanic.groupby("pclass").fare.mean()

titanic.survived.sum()

titanic.survived.mean()

titanic.groupby("sex").survived.mean()

titanic.groupby("pclass").survived.mean()

titanic["ad_chi"] = "adult"

titanic.loc[titanic.age < 18, "ad_chi"] = "child"

titanic.head(20)

titanic.ad_chi.value_counts()

titanic.groupby("ad_chi").survived.mean()

titanic.groupby(["sex", "ad_chi"]).survived.count()

titanic.groupby(["sex", "ad_chi"]).survived.mean().sort_values(ascending = False)

w_and_c_first = titanic.groupby(["sex", "ad_chi"]).survived.mean().sort_values(ascending = False)

w_and_c_first.plot(kind = "bar", figsize = (14,8), fontsize = 14)
plt.xlabel("Groups", fontsize = 13)
plt.ylabel("Survival Rate", fontsize = 13)
plt.title("Titanic Survival Rate by Sex/Age-Groups", fontsize = 16)
plt.show()

### Advanced Aggregation with agg()

import pandas as pd

titanic = pd.read_csv("titanic.csv", usecols = ["survived", "pclass", "sex", "age", "fare"])

titanic.head()

titanic.groupby("sex").mean()

titanic.groupby("sex").sum()

titanic.groupby("sex").agg(["mean", "sum", "min", "max"])

titanic.groupby("sex").agg({"survived": ["sum", "mean"], "pclass": "mean", "age": ["mean", "median"], "fare": "max"})

### GroupBy Aggregation with Relabeling (new in Version 0.25)

import pandas as pd

titanic = pd.read_csv("titanic.csv", usecols = ["survived", "pclass", "sex", "age", "fare"])

titanic.head()

titanic.groupby("sex").survived.mean()

titanic.groupby("sex").agg(survival_rate = ("survived", "mean"))

titanic.groupby("sex").agg({"survived": ["sum", "mean"], "age": ["mean"]})

titanic.groupby("sex").agg(survived_total = ("survived", "sum"), 
                           survival_rate = ("survived", "mean"), mean_age = ("age", "mean"))

### Transformation with transform()

import pandas as pd

titanic = pd.read_csv("titanic.csv")

titanic.head()

titanic.groupby(["sex", "pclass"]).survived.transform("mean")

titanic["group_surv_rate"] = titanic.groupby(["sex", "pclass"]).survived.transform("mean")

titanic.head()

titanic["outliers"] = abs(titanic.survived-titanic.group_surv_rate)

titanic[titanic.outliers > 0.85]

### Replacing NA Values by group-specific Values

import pandas as pd

titanic = pd.read_csv("titanic.csv")

titanic.head(20)

titanic.info()

mean_age = titanic.age.mean()
mean_age

titanic.age.fillna(mean_age)

titanic.groupby(["sex", "pclass"]).age.mean()

titanic["group_mean_age"] = titanic.groupby(["sex", "pclass"]).age.transform("mean")

titanic.head(20)

titanic.age.fillna(titanic.group_mean_age, inplace = True)

titanic.head(20)

titanic.info()

### Generalizing split-apply-combine with apply()

import pandas as pd

titanic = pd.read_csv("titanic.csv", usecols = ["survived", "pclass", "sex", "age", "fare"])

titanic.head()

titanic.groupby("sex").mean()

female_group = list(titanic.groupby("sex"))[0][1]
female_group

female_group.mean(numeric_only=True).astype("float")

def group_mean(group):
    return group.mean(numeric_only=True)

group_mean(female_group)

titanic.groupby("sex").apply(group_mean)

titanic.nlargest(5, "age")

def five_oldest_surv(group):
    return group[group.survived == 1].nlargest(5, "age")

titanic.groupby("sex").apply(five_oldest_surv)

### Hierarchical Indexing (MultiIndex) with Groupby

import pandas as pd

titanic = pd.read_csv("titanic.csv", usecols = ["survived", "pclass", "sex", "age", "fare"])

titanic

summary = titanic.groupby(["sex", "pclass"]).mean()

summary

summary.index

summary.loc[("female", 2), :]

summary.loc[("female", 2), "age"]

summary.swaplevel().sort_index()

summary.reset_index()

### stack() and unstack()

import pandas as pd

summer = pd.read_csv("summer.csv")

summer.head()

medals_by_country = summer.groupby(["Country", "Medal"]).Medal.count()

medals_by_country

medals_by_country.loc[("USA", "Gold")]

medals_by_country.shape

medals_by_country.unstack(level = -1)

medals_by_country = medals_by_country.unstack(level = -1, fill_value= 0)

medals_by_country.head()

medals_by_country.shape

medals_by_country = medals_by_country[["Gold", "Silver", "Bronze"]]

medals_by_country.sort_values(by = ["Gold", "Silver", "Bronze"], ascending = [False, False, False], inplace = True)

medals_by_country.head(10)

import matplotlib.pyplot as plt
plt.style.use("seaborn-v0_8")

medals_by_country.head(10).plot(kind = "bar", figsize = (12,8), fontsize = 13)
plt.xlabel("Country", fontsize = 13)
plt.ylabel("Medals", fontsize = 13)
plt.title("Medals per Country", fontsize = 16)
plt.legend(fontsize = 15)
plt.show()

medals_by_country.stack().unstack()
